Newly edited files --
Registration is now connected to the database, and DBoperations has most viable methods, it should be noted that there are still some bugs that disallow the creation of appointments, but I'll try to fix that as best as I can.  Put these files in your KBS_Webpages directory and overwrite the old files, as well, I've included my web config, it will allow for access to the database.