Add files via upload
These files will have to be linked up to the others - mainly the InsertAdvisorTimes connected to the advisor portal - since we originally talked about importing it into the full on module of admin monitoring page I didn't format it, but that can be done if need be. 
Otherwise, appointments works with a gridlist, the delete and select function work, and the registration is now completely working! It doesn't redirect to anything yet, but it works.